create useSignUp hook to handle sign up functionlaity
